[apps/i18n] Added country support
Added a list of supported countries to I18n, and created messages for
each country's names in each language.

+++ escher/include/escher/i18n.h
@@ -5,7 +5,7 @@
 
 namespace I18n {
   enum class Message : uint16_t;
-  enum class Language : uint16_t;
+  enum class Language : uint8_t;
   const char * translate(Message m);
   int numberOfLanguages();
 }
